"3:55pm. https://www.reddit.com/r/ProgrammingLanguages/comments/e0yf1s/hilbert_bourbaki_and_the_scorning_of_logic/
Finally, here is the thread. As I haven't committed up to now, let me do that here.

Amazing how long this took me.

5pm. Had to take a break and do the chores while I was at it.

Damn, I really do feel tired at this point. But let me get that C# Metamath source and I will start studying it.

5:05pm. http://us.metamath.org/downloads/Verifier.cs

Parsing seems to be half the complexity of this. Let me open a project.

5:20pm. http://us.metamath.org/downloads/mmamm.m

Here is the Mathematica version. This is unreadable as it gets.

http://www.cas.mcmaster.ca/sqrl/papers/SQRLreport18_rev2.pdf
The seven virtues of simple type theory

I kind of want to read this.

https://arxiv.org/abs/1412.8091
Conversion of HOL Light proofs into Metamath

So it is possible to do this automatically.

Mario Carneiro seems to have done a lot work around Metamath.

5:30pm. Ok, I've opened the project, got the C# version. All I need to do now is start studying it.

That having said, it is 5:30pm. Let me just read that simple TT paper and I will call it a day.

Tomorrow I will start playing around with it. Though, I'll have to go outside the house for a while.

"Mathematicians are committed to rigorous reasoning, but they usually shy away from formal logic. However, when mathematicians really need a formal logic—e.g., to teach their students the rules of quantification, to pin down exactly what a “property” is, or to formalize set theory—they almost invariably choose some form of first-order logic. Taking the lead of mathematicians, scientists and engineers usually choose first-order logic whenever they need a formal logic to express mathematical models precisely or to study the logical consequences of theories and specifications carefully. In science and engineering as well as in mathematics, first-order logic reigns supreme!

A formal logic can be a theoretical tool for studying rigorous reasoning and a practical tool for performing rigorous reasoning. Today mathematicians sometimes use formal logics as theoretical tools, but they very rarely use them as practical tools. Scientists and engineers do sometimes use formal logics as practical tools—but not often. Tomorrow things will be different. In the future, mathematicians, scientists, and engineers will routinely employ computer systems—the successors of today’s computer algebra systems and computer theorem proving systems—that mechanize various aspects of mathematical reasoning. In order to be trustworthy, these systems will inevitably be based on formal logics. Also, both theoreticians and practitioners will utilize huge digital libraries of mathematics. The mathematical knowledge in these libraries will be formalized, organized,certified, searched, and retrieved with the help of formal logics."

6:25pm. Done with lunch. I am tuning out here.

As it turns out, I could not start work on studying Metamath source today, but I will be sure to do that tomorrow. After I am done translating it to F#, I will have firm basis from which to move forward.

6:30pm. With the way things are developing, I might end up studying Metamath for a good long while. I might even end up studying HOL at this rate.

Whatever the case is, I am going to have to master classical math in order to get both the confidence and the experience required to tackle the kind of math that I want to really do.

As expected, if you really want something, maybe the best thing to do is to actually take some time and implement it.

I need to get good at this. At the very least, once all is said and done, if I still cannot grasp the papers like the ones like Blackwell then it will be the fault of whoever wrote them. Because my mathematical understanding at that point is not something that will be in question.

That way of thinking is too defeatist though for my tasted.

It is better to go forward with the intention of getting it all!"