"9:45am. Going to bed early did not save me from waking up late it seems.
10:15am. Ok, I've made a reply in that thread. Made that dentist's appointment for Thursday. Only the exercise routine is left. Let me do it now.

10:25am. Up to 18 push-ups. Maybe there is something to this exercise thing. I do not even feel ill today from it.

Now, since it is Monday, let me go for Tog and Satanophany (if the raw is out) and then I will start studying Metamath.

10:50am. Let me finally start.

11:50am. Had to take a break for a bit. Let me take a real break instead. The code is actually rather complex. I am going to take a look at the Haskell version as well.

Even though it is complex, from what I can tell, the first 400 lines are just parsing. Meh.

12pm. http://home.solcon.nl/mklooster/repos/hmm/HmmImpl.hs

This Haskell code is really not inspiring me. It feels too complex.

Ah, screw it. I have no choice but to start with the parser. I'll bring in Fparsec and just start writing things according to spec with the aid of the Python and the C# version.

12:05pm. I'll forget about the semantics and just focus on getting the basic AST processed. After that I'll figure out the actual typechecking and evaluation.

None of this should be too hard once I know how it goes. As a bonus, it should teach me a little about how to do unification."