"3:05pm. I am back. It turns out my teeth are fine - I seem to be a bit too good at brushing them. I need to chill for a bit before I resume. I am feeling jet lag.
My will is weak. I fear the outside world. I just acquire phobias due to stress, and yet I must continue fighting.

3:10pm. Faith is important. I had faith that I would acquire something after I create Spiral, but that has not manifested itself. 2019 is such an aimless year.

After trying these proofs for real, I understand now. I am probably not going to be able to close the gap in math proficiency between me and people like Blackwell and Hart. Having them as my advisors and being able to guide me step by step through the proofs would be such a huge advantage that I am never going to be able to bridge it between me and those that have it. It is not a realistic goal unless you are some kind of genius.

If past is any indication, I am going to learn Metamath, learn set theory and then still be confused as to how to do these proofs. Yet I must do it, even if I am going to get beaten by various other kinds of gaps. I need to put the odds just a little bit more in my favor.

3:20pm. I've come to accept - the math that is not constructive, and that is most of it, is logically inconsistent. And because I do not have the aptitude for that sort of inconsistent math, I am going to have to swim against the current.

Basically I've failed. Spiral was the only thing I could really do. And Spiral was a bit like a desperation move.

The thing that always made sense to do is the one thing that I could not do - figure out the algorithms that I need.

That is the one thing that I should have done. But I cannot do it. I might as well be doing magic.

All I can do is acquire new tastes. The only thing I can envision making me better at this point in constructive math and dependently typed programming.

3:25pm. I cannot learn false math, so I will learn true math. When computer science displaces classical mathematics on the throne of science, then I will be able to win.

It is disappointing. I thought that using the computers I might be able to get good at proving the kinds of theorems I've seen in ML papers, but if anything, learning formal theorem proving is pushing me further away from that. I thought that with effort I might be able to get on the level of pro researchers, but I understand now. I will never be a part of that clique.

I will be their nemesis through and through.

3:30pm. It seems that as long as I am human I am destined to be a loser. At least my programming is good enough for some low end job if it comes to worst. I mean if math of all things is dominated by cliques, then I can't help but wonder if egalitarianism of programming is just my imagination.

But I do not want to leave the honor - and the burden of chasing after the Singularity to anybody else.

3:35pm. But that sort of goal is too far for me.

Since I do not have anything better, the least I can do is cover my holes in the area of mathematics. This time properly. This time justly. What logicians of the 20th century failed, I will do myself - with the help of computers and their science.

3:40pm. Let me catch up with Goblin Slayer and then I will resume. I should take it easy.

Even if I am not doing work, the times that I am relaxing are prime opportunities to establish my motivation. I need to set in stone why constructive mathematics is really worth doing. It need to set in stone, why redoing math from the ground up is worth doing.

Because even if I cannot deal with those papers, the path to figuring out the algorithms definitely should be there."