"8:10am. Ok, enough slacking. Enough reading social media. What a waste of time.
It is time to start. I went to bed really late (I ended up watching Chouyou) and I ended up waking what must have been a few hours ago, but strangely enough, I am not too tired.

First of all, let me do those exercises.

Then I need to deal with the review. Then comes programming.

8:25am. Just catching my breath. I really should just stop this. Let me do the review. This time I'll aim to do it as briefly as possible. I am sort of starting the beginning of the end as far as this quest is concerned. Once I get actual power I will go back to my earlier personality and stop posting my every thought.

---

The highlight of past month is me figuring out theorem 3 from the original CFR paper and [implementing](https://github.com/mrakgr/CFR-In-Fsharp/blob/a2d06683e93a7dd8e9eae15bccf560c933fd53b2/CFR%20In%20Fsharp/Theorem%203/proofs.fsx) it via randomized testing. Right now I am working on a [Metamath verifier](https://github.com/mrakgr/Metamath-in-Fsharp) as a part of my to learning of [set theory](https://www.youtube.com/watch?v=U75S_ZvnWNk).

Due to various experiences over the past year, I've come to a decision - I am going to stop thinking about mathematics as a hard science and mentally move it closer to fields such as gender studies. For the sake of whatever is left of my mental sanity. It is clear to me that at this moment, computer science is really the only field that I know of that actually has a notion of rigor, and a sensible definition of what a proof is. Math certainly does not.

I've been struggling with math in general and have been bemoaning my fate. I mean, in late April I've started going through Software Foundations and learning Coq in order to get familiar with what theorem proving actually is, but to my great sadness the kinds of proofs that I wanted to understand have very little to do with how things are done in Coq. To start with, I wanted to very basic things like how real numbers work and how to integrate a probability distribution from minus to plus infinity. Because I was too dumb to figure that out on my own how to deal with the infinities and expected that surely the entire community of mathematics would know something like that given that it is the bedrock of everything. How naive I was.

There is no way to reconcile how Coq and Agda do proofs and some of the mainstream 'proofs' that I've seen. And the type theoretic languages are right. Meaning unless math is really content to become gender studies, it is going to have to change.

Seen from one side, it is not like I can really do anything about it apart from bitching. I can't really do anything about the numerous disadvantages of the present state of affairs. It is not just the shaky notion of what a proof is, but every field seems to develop its own domain specific language that sort of looks like the regular thing, but is impenetrable to outsiders, and even to specialists from other fields. I can't change this situation and talking about it won't change anything.

But on the other hand, currently the arms race to trigger the technological singularity is ongoing. The tribulations affecting me must also apply to other players in the game. The present goofy state of things might prevent me from becoming a 'serious' researcher, but it might also be a huge opportunity in the long term if I can clear the prerequisites.

It is not at all according to plan - I was absolutely sure that I would be able to simply master the foundations of mathematics and move on to the frontier. How could have I ever expected that foundations of math would the frontier? I mean they had over 2000 years to think this through, didn't they? At this point though, I have enough evidence to accept that the reason why I can't trace the path all the way to the root is not my fault. It is not my fault that I can't fit the pieces together. I cannot keep lumbering forward in disbelief like a zombie.

Much like I did in the past with [Spiral](https://github.com/mrakgr/The-Spiral-Language/tree/v0.1), if I want to get closer to figuring out the secrets of artificial intelligence, I am going to have to innovate in very basic areas, and in places that were considered closed. I won't be able to find the answers if I just look for long enough. I will have to create them myself.

It was too much to expect that I would be able to become an algorithms jockey. I will accept the call and do what is needed in the present in order to bring about the future that I want.

---

9:25pm. Yeah, this is right. This is the correct story. I've buzzing around this direction. The hints were there. Now it is time to come to a conclusion.

I will reject reality the others impose on me and create my better own. It will take me a while, but I have the power to do it.

The machines will help me in this. I was the machines that started all this and they will end it.

One day the human era will end.

9:35am. Is the thread up? Not yet.

Well, it is fine. Let me work on those parsers.

It is time to finish the verifier."