"5:20pm. "Open up any ML paper and you will see functions like `exp` and `log` used in proofs. In addition to that, I've yet to see a paper that would hesitate to use reals. And the probability distributions - let me take the Gaussian as example uses `π`, `exp`, `sqrt` and has a range that goes from -infinity to infinity. In fact actually integrating it requires doing it over that range. To me this is highly dubious from a computational perspective and I've found the proofs of that too creative for my taste.
By now I have some experience in formal theorem proving. I've gone through most of Software Foundations in Coq and the first volume for Agda, and have played around with both Lean and Agda. But it hasn't really scratched my itch - because even with all my knowledge it feels like the actual mathematics that I need to do machine learning are nowhere to be found.

I've had a certain bit of insight about this. A few months ago, I've [speculated](https://github.com/mrakgr/The-Spiral-Language/blob/1001c871ef71311256bd221dd46337dd8a0cc2c7/Learning/Papers/principle%20of%20reversibility.md) that computational reversibility is one of the fundamental principles of learning and while looking for counterpoints I've realized the very first example - the one with the square root, was very unfortunately (or fortunately depending on you looked at it,) picked. Having gone through fairly mainstream schooling, I had not at all realized at the time I wrote that draft that [square root is in fact not defined](https://www.youtube.com/watch?v=REeaT2mWj6Y) for many rationals.

This is significant because it means that it is only possible to approximate the square root, and that in turns is significant because optimization dynamics appear in mathematics much earlier than is appreciated. It is not the optimization of neural nets, or some other ML algorithm where difficulties first appear, but rather with these simple transcendental functions.

Even if I were to use reals and non-constructive mathematics, it is seems obvious to me that this is only a hack. Even if you use *reals*, you are just going to run into the same issues again on a much larger scale when you try to reason about ML.

So I've been thinking about a new approach that would be closer to how more realistic reasoning about functions that can only improve, but possibly never converge would look like.

```
open import Data.Product
open import Data.Nat using (ℕ;suc;zero)
open import Data.Rational
open import Level using (Level)
open import Relation.Nullary
open import Data.Bool using (Bool;false;true)

∣_∣ : ℚ → ℚ
∣ x ∣ with x <? 0ℚ
∣ x ∣ | no ¬p = x
∣ x ∣ | yes p = - x

-- Some function f converges to g, as its precision p increases.
_~>_ : ∀ (f : ℕ → ℚ) → (g : ℕ → ℚ) → Set
f ~> g = ∀ (p : ℕ) → ∃[ p' ] (∣ f p - g p' ∣ ≥ ∣ f (suc p) - g p' ∣)
```

While it might not be possible to implement the idealized `sqrt` directly as in math textbooks, you could give it an extra argument `p` that acts as fuel for how detailed the approximation should be. The idea of an encoding quality is ubiquitous in computing, and it might be a good idea to start with it on the smallest of scales.

So while with rationals `sqrt(x) * sqrt(x)` might never be exactly equal to `x`, it might possible to prove that `sqrt(x) * sqrt(x)` converges to `x` using the above definition of `~>`. I haven't tried it yet though, but I feel it should be doable. And if that turns out to be the case, it might open the way to reasoning about transcendentals in a constructive fashion. To me most of the interesting problems in computer science are optimization problems, and this way of thinking where precision is made explicit really chimes with me.

I am going to try to make something of this, but before I try that let me ask - has what I've illustrated here been done before?

I mean, what I am proposing here is hardly complicated - this sort of relation is the first thing somebody would think of so I am guessing it occurred to somebody else before. But my math sophistication is not that high so I've had to have overlooked it somewhere. Or maybe there is no merit to the idea. Any reference to this would be appreciated."

Let me give this a try. It is really taking me a long time to gather my courage. I need to do this last check before I just dive in and start doing it.

5:30pm. It is issue 990 on the `agda-stdlib` repo. I'll se how it goes.

5:40pm. What should I do now? I certainly do not feel like doing anything in Agda.

I suppose I could figure out how to use Git from Emacs, but I really do not feel like doing that now. There is a general sense that this should be done as a part of an Agda task.

Let me watch a vid by Wildberger and then I am done.

Right now my next move hinges on the kind of replies I will get to the post I've made on the agda-stdlib repo. This late in the day I do not have any desire to start anything new anyway."