"10:05am. https://stackoverflow.com/questions/59263530/how-to-prove-that-the-halving-function-over-positive-rationals-always-has-an-exi
Well, here is the question. For the time being, let me look into how well funded recursion works in Agda.

https://galois.com/blog/2010/06/tech-talk-introducing-well-founded-recursion/

Agh. The links are broken, forget this one.

https://stackoverflow.com/questions/19642921/assisting-agdas-termination-checker

11:15am. I am still thinking about this. Well funded recursion is not the problem here. There are all sorts of subtle issues. With rationals just because a relation like `... < a < b < c` might exist does not mean that the recursion will terminate as with naturals.

Even if at every step there is improvement does not mean I will get what I want. And the termination checker is complaining about that.

I am running into a variant of the continuum problem here. And I have absolutely no idea what to do. There are all sorts of subtle issues that I have no idea how to capture in logic.

...No way going around it. Let me go into that thread and show what I've done so far.

If I cannot get an answer to the SO question, this will be a long project. I'll continue thinking about it into the future, but in the interim I will fully switch to randomized testing. I'll have to study rational math either way.

11:25am. Let me do that thread and then I will have breakfast."