"9:50am. I am up. No replies to the halving question or any comments added to the thread I've opened.
Right now I feel like I have some freedom to chose my next step. Long term I am going to have to get familiar with integers and rationals to the point of being able to reconstruct the library modules on my own and filling in those postulates.

But that is not too important at the moment.

Today, let me chase the answer to the question a bit. I'll go to the Agda mailing list and push the question under their nose. Or the maybe a separate thread on the agda-stdlib repo.

...Let me do a separate thread. I do not want to register for the list.

...Uhm, the thread I opened is right at the top of the repo so I do not want to repost so soon...and I need manual approval to register to the Agda mailing list.

10:05am. I guess there is nothing to it. I'll have to wait to get approval before asking that question on the Agda mailing list.

At this point I am decently sure the halving question is a pretty difficult one. I doubt nobody has seen it.

Forget this issue for the time being.

What exactly should I do next? Well, since dependently typed langs are starting to run into potentially unresolvable termination issues with the kinds of proofs that I want to do...that just further puts randomized testing to the fore.

10:10am. I suppose I could take a look at some more Metamath proofs, but otherwise how I go back to the paper by Hart and see if I can push through it?

Long term, my plan for figuring out ML is to go from rationals to ranges to probability distros and NNs, but right now, I am still in the middle of studying game theory.

10:15am. Oh, I got access.

https://lists.chalmers.se/pipermail/agda/2019/010756.html
[Agda] Do you use postulates extensively?

Right now I am looking at random threads.

10:25am.

```
https://stackoverflow.com/questions/59263530/how-to-prove-that-the-halving-function-over-positive-rationals-always-has-an-exi

Sorry for being pushy, but lately I've gotten an interest in rational math and I really want some input from the experts here regarding how to prove termination of the above. Implementing the function was easy, but proving the termination seems impossible so at the very least I need to understand whether that impression is right or not. Silence as the reply just has too many possible interpretations to deal with.
```

10:30am. Now let me chill.

https://boards.4channel.org/a/thread/196552885/uq-holder-165167-2019-digital-dankeempire

I'll leave this thread for later. Let me just do my morning reading for a little bit and then I will do a bit of work.

11:05am. That was fun. Let me finally start...actually, let me just have breakfast here.

It is not like I have a clear plan of action to be starting right now anyway.

11:10am. Let me take some time to think what my plan should be...

How about I look at Metamath proofs for an hour or so and then get back to the paper by Hart?

What is the problem with that?

The problem is that I really learned very little set theory in the past two weeks.

So set theory! Set theory should be my focus today. That is what I need to study.

Since Metamath proofs are turning out tedious, and the whole thing is not going exactly according to plan, it might be a decent idea to resume those videos.

Either way, I should be prepared for some tedium.

But I need to do this.

...Actually, how about I give Mizar a serious try as well?

11:15am. December should be my set theory month. I need to do this once in my life. Set theory might be a broken system, but it is a broken system everybody else is using. Just like I have with type theory, I need to attain some level of proficiency in it. If Metamath won't do the job for me, I should try Mizar.

What is important?

It is important that I do not just leaf through proofs or watch videos. To get proficiency, I really need to use set theory myself.

11:20am. Then isn't it a given that I have to get Mizar and do some programming in it? It is.

Mizar will be my next target then.

But I will take half hour of going through more Metamath proofs to start things off.

11:25am. Ok, with the plan set, let me have breakfast here, do the chores and then I will do my work for the day."