"1:25pm. https://mangadex.org/title/43308/youkai-gurentai
The first chapter of this was interesting enough.

But now let me resume. I am actually interested in what the book has to say.

1:55pm. https://stackoverflow.com/questions/59263530/how-to-prove-that-the-halving-function-over-positive-rationals-always-has-an-exi/59304233#59304233

Got an answer. I think I see it now.

Let me stop the book for a bit, so I can implement this.

2:10pm. Actually, there is too much work here right now. But the steps he suggests are sound.

2:20pm. Let me go back to the book by Bishop. I will leave the concrete proofs when I really dive into integers and rationals in Agda.

2:45pm. At chapter two. Thankfully the real numbers are early in, so I won't have to wade through a lot of muck to see if the rest of the book is worth reading.

"The next three sections are devoted to constructing the real numbers, as certain Cauchy sequences of rational numbers, and investigating their order and arithmetic."

Cauchy sequences. Yuck.

"Section 1 establishes some conventions regarding sets and functions. "

And apparently this book will be about set theory.

...Well, I'll give a few dozen pages a try.

3:05pm. 26/380.

```
Let a be any rational number. We define
(2.4)
    (a) x + у = {x2n + 2/2n}2U
    (b) ХУ = {Х2кпУ2кп} n~l
    (c) max {x,y} = {max {xnyyn}}^
    (d) -χ = {-xn}?-i
    (e) α* = {α, α, α, . . .}
```

This is just Cauchy sequence crap. The paste is bad which makes it hard to see, but there are infinite sets in use here.

3:15pm. Let me take a little break here to read Omega. I won't be continuing with this book. This is not real constructive mathematics. Even though he talks a good story in the beginning, I guess it was too much to expect for those in the 60s to do serious algorithmics. I guess a language on par with Agda really is needed to even start thinking about doing things in the correct direction.

What a waste of time.

Let me read Omega, then I will check out that other paper, and then I will get back to the CFR papers as per plan.

3:50pm. Seriously, that book in the introduction pushed all the right buttons, but then come infinite sequences and sets. I am not expecting much from the probability theory book at this point.

"Introduction

The present work on probability theory is an outgrowth of the constructive analysis in
[Bishop 1967] and [Bishop and Bridges 1985].

Perhaps the simplest explanation of constructive mathematics is by way of focusing
on the following two commonly used theorems. The first, the principle of finite search,
states that, given a finite sequence of 0-or-1 integers, either all members of the sequence
are equal to 0, or there exists a member which is equal to 1. We use this theorem without
hesitation because, given the finite sequence, a finite search would determine the result.

The second theorem, which we may call the principle of infinite search, states that,
given an infinite sequence of 0-or-1 integers, either all members of the sequence are
equal to 0, or there exists a member which is equal to 1. The name “infinite search”
is perhaps unfair, but it brings into sharp focus that the computational meaning of
this theorem is not clear. The theorem is tantamount to an infinite loop in computer
programming."

At least this guy is honest.

Well, I'll give the paper a try for a bit.

4:20pm. Ok, enough of this. Pretty much everything in this paper from the get go is complicated set theory, or infinite this or infinite that.

I really misjudged constructivism. I am not sure what it is supposed to be at all now, but whatever it is, I do not want it.

There is no doubt, that I do need to master the things found in these books in general, but I will do it gradually as I build up my foundations. And I will do so with an eye towards concrete algorithms.

4:25pm. Maybe trying to study these books was an excuse.

As I said, I slept really poorly tonight, and I am dead tired. Even if I try the paper by Hart, I won't be able to get any flow going the way I am now. But let me give it a try. I need to do at least something to get closer to my goal.

Things are going to get serious starting from here. I've found my focus, and I am going to use it to great effect.

The things you can do in Agda goes way up once you recognize that the termination checker is a limitation.

Ultimately, proofs are computer programs, and the more power you can focus into a single strike, the further you can get. That is the ability I need to cultivate.

4:40pm. I will never accept axioms or postulates as valid proofs, but ignoring the termination checker would be acceptable on a case by case basis. Because a proof might be valid even if the termination checker does not recognize it as such.

I really love this sort of thing.

```
Let ## N, S , u . . be a finite N-person game in strategic normal Ž . i$ N i$ N form: N is the set of players, Si is the set of strategies of player i, and ui : Ł Si !! is player i’s payoff function.
```

The paste is bad as usual. Let me start formalizing stuff from 3/32.

4:50pm. Let me just break the ice today. Like I wanted to with CFR, what I need to do here is start with the basic definition of a game.

5:05pm.

```
Game = Σ[ n ∈ ℕ ] (Fin n → ℚ)
```

I am having a difficult time figuring out how to put in the definition of a strategy, but despite that I am getting a decent amount of flow going. I am in the zone.

5:20pm.

```
Game = Σ[ n ∈ ℕ ] Σ[ Strategy ∈ (Fin n → Set) ] (Fin n → (∀ (i : Fin n) → Strategy i) → ℚ)
```

I think this is the definition I am looking for. It is great.

```
Game = Σ[ n ∈ ℕ ] Σ[ Strategy ∈ (Fin n → Set) ] (Fin n → (∀ i → Strategy i) → ℚ)
```

Let me remove a needless type annotation.

5:25pm. Rather than a function it would be better if there was Vec whose type was indexed, but I'll leave that aside for now. Efficiency considerations are not even close to my main worry right now.

At any rate, it is great that now I have my first type theoretic interpretation from the paper by Hart.

5:30pm.

```
A probability distribution $ on S is a correlated equilibrium of #
if, for every i$N, every j$Si and every k$Si we have
```

Now comes a crazy definition.

5:40pm. I think I am getting some insight.

```
Γ = Σ[ n ∈ ℕ ] Σ[ Strategy ∈ (Fin n → Set) ] (Fin n → (∀ i → Strategy i) → ℚ)
```

...

Ok, I think I understand why this definition confused me so much last time.

Imagine I had `n : Nat` and then the next sentence said, `sum over n in Nat`. At first he is talking about types, but then it outright treats strategy as a set to sum over.

6:05pm. I've been deep in thought all this time. I am still thinking about it, but regardless, I think I completely understand the definition now.

That having said.

```
Γ = Σ[ n ∈ ℕ ] Σ[ Strategy ∈ (Fin n → Set) ] (Fin n → (∀ i → Strategy i) → ℚ)
```

I can't just iterate over an arbitrary set of strategies like in set theory. I need to find a different formulation in type theory.

6:15pm.

```
Strategy : ℕ → Set
Strategy n = Fin n → ℕ

Γ = Σ[ n ∈ ℕ ] Σ[ S ∈ Strategy n ] (Fin n → (∀ i → Fin (S i)) → ℚ)
```

Closer, but not enough.

6:30pm.

```
_∈_ : ∀ {A} → A → List A → Set
k ∈ l = Any (k ≡_) l

Γ = Σ[ n ∈ ℕ ] Σ[ Strategy ∈ (Fin n → Set) ] Σ[ S ∈ (∀ i → List (Strategy i)) ] (Fin n → (∀ i x → x ∈ S i) → ℚ)
```

This really makes things quite complicated. Really, the easiest way to do this would be an idea I had before.

6:35pm.

```
Γ = Σ[ n ∈ ℕ ] Σ[ Strategy ∈ (Fin n → ℕ) ] (Fin n → (∀ i → Fin (Strategy i)) → ℚ)
```

Annoyingly, this really does meet all my requirements. Using Nats everywhere is ugly, but this really would generalize to any kind of game.

I think I will stop here for the day."