"11:25am. I actually got up 5m ago. I think this is the very latest I've ever gotten up in not-so recent memory. I slept spectacularly well though, and I am extremely well rested. Wonderful.
Since it is this late, I might as well have breakfast now.

11:55am. Done with breakfast. Let me do the chores and then I will start.

12:25pm. Let me finally resume where I left off from yesterday.

At the end of the last day, I defined the game as ...

```
Γ = Σ[ n ∈ ℕ ] Σ[ Strategy ∈ (Fin n → ℕ) ] (Fin n → (∀ i → Fin (Strategy i)) → ℚ)
```

For simplicity I'll turn this into a record.

The next thing on the list of things to do would be the definition of the correlated equilibrium. This gave me a lot of trouble weeks ago, but now that I am starting to awaken to my real power I am making headway on this.

Let me open the paper. Let me rewrite the definition by hand.

```
A probability distribution fi on S is a correlated equilibrium of Gamma if for every i in N , every j in S^i and every k in S^i we have sum S (fun s (s^i = j) -> fi(s) * (u^i(k,s^-i) - u^i(s))) <= 0
```

Why did this give me so much trouble last time.

```
u^i(k,s^-i) - u^i(s)
```

It is this thing here. I tried imagining this as a fold, but no matter how I tried wrapping my mind around this, it seems obvious that you can't just put `(k,s^-i)` in this order as if it were a tuple because that would wreck the order of the arguments to `u`.

And indeed that is the case. This is modern mathematical rigour in action. These people are all bark and no bite.

But just for the sake of the argument imagine that `u(s^i,s^-i)` is valid. Then...

```
sum S (fun s (s^i = j) -> fi(s) * (u^i(k,s^-i) - u^i(s))) <= 0
= sum S (fun s (s^i = j) -> fi(s) * (u^i(k,s^-i) - u^i(s^i,s^-i))) <= 0
= sum S (fun s -> fi(s) * (u^i(k,s^-i) - u^i(j,s^-i))) <= 0
```

And now it becomes obvious what the definition is doing. It is iterating over the strategies for the opponent.

12:35pm. The thing that is really giving me trouble is quite simple. It is not that I do not understand set theory. It is that I am mentally failing to compile it and unsure how to proceed.

The solution is not to take math so literally it seems. It is a skill I am going to have to have to learn.

12:45pm. Agh, just writing the above has made me switch gears into rant mode. Let me take just a short break and then I will resume.

12:50pm. Ok, focus me. What should I do next?

"Also, the set of correlated equilibria is nonempty, closed and convex, and even in simple games e.g., Ž ‘‘chicken’’ it may include distributions that are not in the convex hull of the . Nash equilibrium distributions"

I have no idea what this means, but let me leave it for later. Just put that into the fiction category for now.

12:55pm. What I will do first is make that definition of the correlated equilibrium.

1:20pm. Yes, I could translate the paper by Hart into type theory. Or I could argue on /sci/.

">>11222458
>completeness theorem
I am not familiar with those, but...
>unless you accept that "E->F" is not interpreted as F^E but a subset of it, or a variant like a third set M(E,F) and a dedicated "apply" map a_{E,F}: M(E,F)xE -> F )
This is related to how types are like a specification or a program right? So if you have a Nat, in type theory you cannot directly iterate over every possible element of that type, but in set theory you can over the set Nat.

I kind of see this as an advantage though. Yes, it is true that in type theory you can do less out of the box as a consequence of the operations themselves being more primitive. But on the other hand, sets themselves are too complex as data structures to be the foundation. Before you can talk about sets you really should talk about lists, and before you talk about lists you should talk about natural numbers.

Type theory gets the natural order of concepts right, and the fact that it can be compiled and used on a computer is a great argument in its favor as a foundation of mathematics.

Set theory really only has an advantage over type theory if you are doing things on pen and paper and you want to cram down the size of the proof. Then things being less primitive is good. But computers exist in this era and math should go the way of programming."

I wrote this for once. My sense is that I am not really completely ready to argue with these guys just yet as my math knowledge is too low. And the angle that guy argued against me is new.

But oh well. He might say something smart in reply and I will learn from that.

1:25pm. Focus me, focus. I can't tell myself not to waste time, since translating these paper is only an exercise, but this exercise is something I must go through if I am to get better. Hart put a lot of effort into writing this paper and I will get everything I can from it.

```
record Γ : Set where
  field
    n : ℕ
    S : Fin n → ℕ
    u : (∀ i → Fin (S i)) → Fin n → ℚ

correlated-equilibrium : (ε : ℚ) → (γ : Γ) → Set
correlated-equilibrium ε γ = sum γ ≤ ε where
  sum : Γ →
```

What do I write here?

1:45pm.

```
record Γ : Set where
  field
    n : ℕ
    S : Fin n → ℕ
    u : (∀ i → Fin (S i)) → Fin n → ℚ
open Γ

correlated-equilibrium : (ε : ℚ) → (γ : Γ) → Set
correlated-equilibrium ε γ = ∀ ψ → sum' γ ψ ℚ.≤ ε where
  sum' : (γ : Γ) → (ψ : ∀ (i : Fin (n γ)) → Fin (S γ i) → ℚ) → ℚ
```

The issue with this is that it is complex. All those terms definitely feel like cruft over the straightforward set theoretic formulation. And I have not even states that psi should be a probability distribution.

2:10pm.

```
postulate sum-over : ∀ a (b : Fin a → ℕ) → (f : ∀ (i : Fin a) → Fin (b i) → ℚ) → ℚ
is-prob-dist = ∀ a b f → sum-over a b f ≡ 1ℚ × (∀ i j → 0ℚ ℚ.≤ f i j)
is-prob-dist-over-S = ∀ γ f → is-prob-dist (n γ) (S γ) f
```

Damn it, I am bogged down again over this.

2:25pm.

```
open import Data.List as List
open import Data.Vec as Vec
open import Data.Nat as ℕ
open import Data.Fin
open import Data.Integer as ℤ
open import Data.Rational as ℚ
open import Data.Product
open import Relation.Binary.PropositionalEquality
open import Relation.Nullary
open import Data.List.Relation.Unary.Any

record Γ : Set where
  field
    n : ℕ
    S : Fin n → ℕ
    u : (∀ i → Fin (S i)) → Fin n → ℚ
open Γ

postulate sum-over : ∀ a (b : Fin a → ℕ) → (f : ∀ (i : Fin a) → Fin (b i) → ℚ) → ℚ
is-2d-% : ∀ (a : ℕ) → (b : Fin a → ℕ) → (f : ∀ (i : Fin a) → Fin (b i) → ℚ) → Set
is-2d-% a b f = sum-over a b f ≡ 1ℚ × (∀ i j → 0ℚ ℚ.≤ f i j)

correlated-equilibrium : (ε : ℚ) → (γ : Γ) → Set
correlated-equilibrium ε γ = ∀ ψ → sum' γ ψ ℚ.≤ ε where
  sum' : (γ : Γ) → (ψ : ∃[ f ] (is-2d-% (n γ) (S γ) f)) → ℚ
  sum' γ ψ = {!!}
```

This is not working for me. In theory it is doable, but iterating over the functions would just be too much work. I am tempted to just switch to F# and randomized testing at this point.

This won't stand.

Let me take it from the top, this time with `Vec`s.

If that fails as well, I really will go to randomized testing despite my boisterious praise of type theory."